+++ src/main/java/gregapi/tileentity/tank/TileEntityBase09FluidContainerSync.java
@@ -34,18 +34,18 @@ public abstract class TileEntityBase09FluidContainerSync extends TileEntityBase0
 	
 	@Override
 	public boolean onTickCheck(long aTimer) {
-		return super.onTickCheck(aTimer) || (short)FL.id_(mTank) != oDisplay;
+		return super.onTickCheck(aTimer) || FL.id_(mTank) != oDisplay;
 	}
 	
 	@Override
 	public void onTickResetChecks(long aTimer, boolean aIsServerSide) {
 		super.onTickResetChecks(aTimer, aIsServerSide);
-		oDisplay = (short)FL.id_(mTank);
+		oDisplay = FL.id_(mTank);
 	}
 	
 	@Override
 	public IPacket getClientDataPacket(boolean aSendAll) {
-		short tDisplay = (short)FL.id_(mTank);
+		short tDisplay = FL.id_(mTank);
 		if (aSendAll) return getClientDataPacketByteArray(aSendAll, UT.Code.toByteS(tDisplay, 0), UT.Code.toByteS(tDisplay, 1), (byte)UT.Code.getR(mRGBa), (byte)UT.Code.getG(mRGBa), (byte)UT.Code.getB(mRGBa));
 		return getClientDataPacketByteArray(aSendAll, UT.Code.toByteS(tDisplay, 0), UT.Code.toByteS(tDisplay, 1));
 	}
